% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circumstance.R
\name{circumstance}
\alias{circumstance}
\title{circumstance requires pomp}
\usage{
circumstance()
}
\value{
None
}
\description{
This function allows one to have pomp and circumstance
}
\examples{
cirsumstance()

}
\keyword{circumstance}
\keyword{pomp, circumstance}
